requirements.txt
numpy>=1.24.0
pandas>=2.0.0
tvdatafeed-enhanced>=2.2.0
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0
websocket-client>=1.6.0
websockets>=14.0
matplotlib>=3.7.0
